// sdk/src/chain/gen/cosmos/tx/v1beta1/tx.ts
export interface SignDoc {
  /**
   * body_bytes is protobuf serialization of a TxBody that matches the
   * representation in TxRaw.
   */
  bodyBytes: Uint8Array;
  /**
   * auth_info_bytes is a protobuf serialization of an AuthInfo that matches the
   * representation in TxRaw.
   */
  authInfoBytes: Uint8Array;
  /**
   * chain_id is the unique identifier of the chain this transaction targets.
   * It prevents signed transactions from being used on another chain by an
   * attacker
   */
  chainId: string;
  /** account_number is the account number of the account in state */
  accountNumber: bigint;
}
// ...
export interface TxBody {
  /**
   * messages is a list of messages to be executed. The required signers of
   * those messages define the number and order of elements in AuthInfo's
   * signer_infos and Tx's signatures. Each required signer address is added to
   * the list only the first time it occurs.
   * By convention, the first required signer (usually from the first message)
   * is referred to as the primary signer and pays the fee for the whole
   * transaction.
   */
  messages: Any[];
  /**
   * memo is any arbitrary note/comment to be added to the transaction.
   * WARNING: in clients, any publicly exposed text should not be called memo,
   * but should be called `note` instead (see
   * https://github.com/cosmos/cosmos-sdk/issues/9122).
   */
  memo: string;
  /**
   * timeout_height is the block height after which this transaction will not
   * be processed by the chain.
   */
  timeoutHeight: bigint;
  /**
   * unordered, when set to true, indicates that the transaction signer(s)
   * intend for the transaction to be evaluated and executed in an un-ordered
   * fashion. Specifically, the account's nonce will NOT be checked or
   * incremented, which allows for fire-and-forget as well as concurrent
   * transaction execution.
   *
   * Note, when set to true, the existing 'timeout_timestamp' value must
   * be set and will be used to correspond to a timestamp in which the transaction is deemed
   * valid.
   *
   * When true, the sequence value MUST be 0, and any transaction with unordered=true and a non-zero sequence value will
   * be rejected.
   * External services that make assumptions about sequence values may need to be updated because of this.
   */
  unordered: boolean;
  /**
   * timeout_timestamp is the block time after which this transaction will not
   * be processed by the chain.
   *
   * Note, if unordered=true this value MUST be set
   * and will act as a short-lived TTL in which the transaction is deemed valid
   * and kept in memory to prevent duplicates.
   */
  timeoutTimestamp:
    | Date
    | undefined;
  /**
   * extension_options are arbitrary options that can be added by chains
   * when the default options are not sufficient. If any of these are present
   * and can't be handled, the transaction will be rejected
   */
  extensionOptions: Any[];
  /**
   * extension_options are arbitrary options that can be added by chains
   * when the default options are not sufficient. If any of these are present
   * and can't be handled, they will be ignored
   */
  nonCriticalExtensionOptions: Any[];
}

/**
 * AuthInfo describes the fee and signer modes that are used to sign a
 * transaction.
 */
export interface AuthInfo {
  /**
   * signer_infos defines the signing modes for the required signers. The number
   * and order of elements must match the required signers from TxBody's
   * messages. The first element is the primary signer and the one which pays
   * the fee.
   */
  signerInfos: SignerInfo[];
  /**
   * Fee is the fee and gas limit for the transaction. The first signer is the
   * primary signer and the one which pays the fee. The fee can be calculated
   * based on the cost of evaluating the body and doing signature verification
   * of the signers. This can be estimated via simulation.
   */
  fee:
    | Fee
    | undefined;
  /**
   * Tip is the optional tip used for transactions fees paid in another denom.
   *
   * This field is ignored if the chain didn't enable tips, i.e. didn't add the
   * `TipDecorator` in its posthandler.
   *
   * @deprecated
   */
  tip: Tip | undefined;
}
// ...
export interface Tip {
  /** amount is the amount of the tip */
  amount: Coin[];
  /** tipper is the address of the account paying for the tip */
  tipper: string;
}
// ...
export const SignDoc: MessageFns<SignDoc> = {
  encode(message: SignDoc, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.bodyBytes.length !== 0) {
      writer.uint32(10).bytes(message.bodyBytes);
    }
    if (message.authInfoBytes.length !== 0) {
      writer.uint32(18).bytes(message.authInfoBytes);
    }
    if (message.chainId !== "") {
      writer.uint32(26).string(message.chainId);
    }
    if (message.accountNumber !== 0n) {
      if (BigInt.asUintN(64, message.accountNumber) !== message.accountNumber) {
        throw new globalThis.Error("value provided for field message.accountNumber of type uint64 too large");
      }
      writer.uint32(32).uint64(message.accountNumber);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignDoc {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const previousRecursionDepth = (reader as any).__tsProtoDecodeDepth ?? 0;
    if (previousRecursionDepth >= 100) {
      throw new globalThis.Error("protobuf decode recursion limit exceeded");
    }
    (reader as any).__tsProtoDecodeDepth = previousRecursionDepth + 1;
    try {
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseSignDoc();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.bodyBytes = reader.bytes();
            continue;
          }
          case 2: {
            if (tag !== 18) {
              break;
            }

            message.authInfoBytes = reader.bytes();
            continue;
          }
          case 3: {
            if (tag !== 26) {
              break;
            }

            message.chainId = reader.string();
            continue;
          }
          case 4: {
            if (tag !== 32) {
              break;
            }

            message.accountNumber = reader.uint64() as bigint;
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    } finally {
      (reader as any).__tsProtoDecodeDepth = previousRecursionDepth;
    }
  },

  fromJSON(object: any): SignDoc {
    return {
      bodyBytes: isSet(object.bodyBytes)
        ? bytesFromBase64(object.bodyBytes)
        : isSet(object.body_bytes)
        ? bytesFromBase64(object.body_bytes)
        : new Uint8Array(0),
      authInfoBytes: isSet(object.authInfoBytes)
        ? bytesFromBase64(object.authInfoBytes)
        : isSet(object.auth_info_bytes)
        ? bytesFromBase64(object.auth_info_bytes)
        : new Uint8Array(0),
      chainId: isSet(object.chainId)
        ? globalThis.String(object.chainId)
        : isSet(object.chain_id)
        ? globalThis.String(object.chain_id)
        : "",
      accountNumber: isSet(object.accountNumber)
        ? BigInt(object.accountNumber)
        : isSet(object.account_number)
        ? BigInt(object.account_number)
        : 0n,
    };
  },

  toJSON(message: SignDoc): unknown {
    const obj: any = {};
    if (message.bodyBytes.length !== 0) {
      obj.bodyBytes = base64FromBytes(message.bodyBytes);
    }
    if (message.authInfoBytes.length !== 0) {
      obj.authInfoBytes = base64FromBytes(message.authInfoBytes);
    }
    if (message.chainId !== "") {
      obj.chainId = message.chainId;
    }
    if (message.accountNumber !== 0n) {
      obj.accountNumber = message.accountNumber.toString();
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SignDoc>, I>>(base?: I): SignDoc {
    return SignDoc.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SignDoc>, I>>(object: I): SignDoc {
    const message = createBaseSignDoc();
    message.bodyBytes = object.bodyBytes ?? new Uint8Array(0);
    message.authInfoBytes = object.authInfoBytes ?? new Uint8Array(0);
    message.chainId = object.chainId ?? "";
    message.accountNumber = (object.accountNumber !== undefined && object.accountNumber !== null)
      ? BigInt(object.accountNumber)
      : 0n;
    return message;
  },
};
// ...
export const AuthInfo: MessageFns<AuthInfo> = {
  encode(message: AuthInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.signerInfos) {
      SignerInfo.encode(v!, writer.uint32(10).fork()).join();
    }
    if (message.fee !== undefined) {
      Fee.encode(message.fee, writer.uint32(18).fork()).join();
    }
    if (message.tip !== undefined) {
      Tip.encode(message.tip, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AuthInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const previousRecursionDepth = (reader as any).__tsProtoDecodeDepth ?? 0;
    if (previousRecursionDepth >= 100) {
      throw new globalThis.Error("protobuf decode recursion limit exceeded");
    }
    (reader as any).__tsProtoDecodeDepth = previousRecursionDepth + 1;
    try {
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseAuthInfo();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.signerInfos.push(SignerInfo.decode(reader, reader.uint32()));
            continue;
          }
          case 2: {
            if (tag !== 18) {
              break;
            }

            message.fee = Fee.decode(reader, reader.uint32());
            continue;
          }
          case 3: {
            if (tag !== 26) {
              break;
            }

            message.tip = Tip.decode(reader, reader.uint32());
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    } finally {
      (reader as any).__tsProtoDecodeDepth = previousRecursionDepth;
    }
  },

  fromJSON(object: any): AuthInfo {
    return {
      signerInfos: globalThis.Array.isArray(object?.signerInfos)
        ? object.signerInfos.map((e: any) => SignerInfo.fromJSON(e))
        : globalThis.Array.isArray(object?.signer_infos)
        ? object.signer_infos.map((e: any) => SignerInfo.fromJSON(e))
        : [],
      fee: isSet(object.fee) ? Fee.fromJSON(object.fee) : undefined,
      tip: isSet(object.tip) ? Tip.fromJSON(object.tip) : undefined,
    };
  },

  toJSON(message: AuthInfo): unknown {
    const obj: any = {};
    if (message.signerInfos?.length) {
      obj.signerInfos = message.signerInfos.map((e) => SignerInfo.toJSON(e));
    }
    if (message.fee !== undefined) {
      obj.fee = Fee.toJSON(message.fee);
    }
    if (message.tip !== undefined) {
      obj.tip = Tip.toJSON(message.tip);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AuthInfo>, I>>(base?: I): AuthInfo {
    return AuthInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AuthInfo>, I>>(object: I): AuthInfo {
    const message = createBaseAuthInfo();
    message.signerInfos = object.signerInfos?.map((e) => SignerInfo.fromPartial(e)) || [];
    message.fee = (object.fee !== undefined && object.fee !== null) ? Fee.fromPartial(object.fee) : undefined;
    message.tip = (object.tip !== undefined && object.tip !== null) ? Tip.fromPartial(object.tip) : undefined;
    return message;
  },
};